review filepaths
