+++ .github/workflows/review_pr.yml
@@ -15,7 +15,7 @@ jobs:
 
     strategy:
       matrix:
-        python-version: ["3.8"] #, "3.9", "3.9-dev"]
+        python-version: ["3.9"]
 
     steps:
       - uses: actions/checkout@v2
